refactor(home): Move home page into components
Literally it looks SO MUCH BETTER NOW TT; Basically all I did was move the large
ass index.tsx file's code to individual components.
