+++ components/homeComponents/HomeNotif.tsx
@@ -1,0 +1,56 @@
+import { useEffect, useState } from 'react';
+import firebase from 'firebase/app';
+import 'firebase/messaging';
+
+export default function HomeNotif() {
+  const [notif, setNotif] = useState(true);
+
+  useEffect(() => {
+    // Set amount of time notification prompt gets displayed before fading out
+    setNotif(checkNotif());
+    setTimeout(fadeOutEffect, 3000);
+  }, []);
+
+  const checkNotif = () => {
+    //pop up visible if user did not enable push notif and browser supports push notif
+    const isSupported =
+      'Notification' in window &&
+      'serviceWorker' in navigator &&
+      'PushManager' in window &&
+      firebase.messaging.isSupported();
+    if (isSupported && Notification.permission !== 'granted') {
+      Notification.requestPermission();
+      return true;
+    }
+    return false;
+  };
+
+  // Fade out notification prompt
+  const fadeOutEffect = () => {
+    var fadeTarget = document.getElementById('popup');
+
+    if (fadeTarget !== undefined && fadeTarget !== null) {
+      var fadeEffect = setInterval(() => {
+        if (!fadeTarget.style.opacity) {
+          fadeTarget.style.opacity = '1';
+        }
+        if (parseFloat(fadeTarget.style.opacity) > 0) {
+          fadeTarget.style.opacity = (parseFloat(fadeTarget.style.opacity) - 0.1).toString();
+        } else {
+          clearInterval(fadeEffect);
+        }
+      }, 100);
+    }
+  };
+
+  return (
+    notif && (
+      <div
+        id="popup"
+        className="fixed z-50 md:translate-x-0 translate-x-1/2 w-[22rem] rounded-md px-4 py-2 top-16 md:right-6 right-1/2 bg-red-200 md:text-base text-sm"
+      >
+        Turn on push notifications to recieve announcements!
+      </div>
+    )
+  );
+}
